OLCS-14786 Fix issue with OrganisationPerson and ApplicationOrganisationPerson e…

…nties referencing a delete Person

See merge request !176

+++ module/Api/src/Domain/CommandHandler/Application/DeletePeople.php
@@ -12,6 +12,7 @@ use Dvsa\Olcs\Api\Domain\CommandHandler\TransactionedInterface;
 use Dvsa\Olcs\Api\Domain\Command\Result;
 use Dvsa\Olcs\Api\Entity\Application\Application as ApplicationEntity;
 use Dvsa\Olcs\Transfer\Command\CommandInterface;
+use Dvsa\Olcs\Api\Entity\Application\ApplicationOrganisationPerson;
 
 /**
  * DeletePeople
@@ -23,6 +24,13 @@ final class DeletePeople extends AbstractCommandHandler implements Transactioned
     protected $repoServiceName = 'Application';
     protected $extraRepos = ['OrganisationPerson', 'ApplicationOrganisationPerson','Person'];
 
+    /**
+     * Handle comment
+     *
+     * @param \Dvsa\Olcs\Transfer\Command\Application\DeletePeople $command Command
+     *
+     * @return Result
+     */
     public function handleCommand(CommandInterface $command)
     {
         /* @var $command \Dvsa\Olcs\Transfer\Command\Licence\DeletePeople */
@@ -52,26 +60,32 @@ final class DeletePeople extends AbstractCommandHandler implements Transactioned
     /**
      * Delete a (Application)OrganisationPerson by using deltas
      *
-     * @param ApplicationEntity $application
+     * @param ApplicationEntity $application Application
      * @param type              $personId    ID of person to delete
      * @param Result            $result      Result object to add messages to
+     *
+     * @return void
      */
     private function deltaDeletePerson(ApplicationEntity $application, $personId, Result $result)
     {
         try {
+            /** @var ApplicationOrganisationPerson $appOrgPerson */
             $appOrgPerson = $this->getRepo('ApplicationOrganisationPerson')
                 ->fetchForApplicationAndPerson($application->getId(), $personId);
 
-            $result->addMessage("ApplicationOrganisationPerson ID {$appOrgPerson->getId()} deleted");
-            $this->getRepo('ApplicationOrganisationPerson')->delete($appOrgPerson);
-            $this->getRepo('Person')->delete($appOrgPerson->getPerson());
+            // It shouldn't be possible to delete a delete delta, but will cause issues
+            if ($appOrgPerson->getAction() !== ApplicationOrganisationPerson::ACTION_DELETE) {
+                $result->addMessage("ApplicationOrganisationPerson ID {$appOrgPerson->getId()} deleted");
+                $this->getRepo('ApplicationOrganisationPerson')->delete($appOrgPerson);
+                $this->getRepo('Person')->delete($appOrgPerson->getPerson());
+            }
         } catch (\Dvsa\Olcs\Api\Domain\Exception\NotFoundException $e) {
-            $appOrgPerson = new \Dvsa\Olcs\Api\Entity\Application\ApplicationOrganisationPerson(
+            $appOrgPerson = new ApplicationOrganisationPerson(
                 $application,
                 $application->getLicence()->getOrganisation(),
                 $this->getRepo()->getReference(\Dvsa\Olcs\Api\Entity\Person\Person::class, $personId)
             );
-            $appOrgPerson->setAction('D');
+            $appOrgPerson->setAction(ApplicationOrganisationPerson::ACTION_DELETE);
 
             $application->addApplicationOrganisationPersons($appOrgPerson);
             $this->getRepo('ApplicationOrganisationPerson')->save($appOrgPerson);
@@ -83,9 +97,11 @@ final class DeletePeople extends AbstractCommandHandler implements Transactioned
     /**
      * Delete an OrganisationPerson
      *
-     * @param ApplicationEntity $application
+     * @param ApplicationEntity $application Application
      * @param type              $personId    ID of person to delete
      * @param Result            $result      Result object to add messages to
+     *
+     * @return void
      */
     private function deleteOrganisationPerson(ApplicationEntity $application, $personId, Result $result)
     {
@@ -93,16 +109,13 @@ final class DeletePeople extends AbstractCommandHandler implements Transactioned
             $application->getLicence()->getOrganisation()->getId(),
             $personId
         );
-        // There should only be one, but just in case iterate them
-        foreach ($organisationPersons as $organisationPerson) {
-            $this->getRepo('OrganisationPerson')->delete($organisationPerson);
-            $result->addMessage("OrganisationPerson ID {$organisationPerson->getId()} deleted");
-        }
 
-        // if no other OrganisationPerson relates to the person ID
-        if (count($this->getRepo('OrganisationPerson')->fetchListForPerson($personId)) === 0) {
-            $this->getRepo('Person')->delete($organisationPerson->getPerson());
-            $result->addMessage("Person ID {$organisationPerson->getPerson()->getId()} deleted");
+        // There should only be one, but just in case iterate them
+        $organisationPersonIds = [];
+        foreach ($organisationPersons as $organisationPerson) {
+            $organisationPersonIds[] = $organisationPerson->getId();
         }
+        $dto = \Dvsa\Olcs\Transfer\Command\OrganisationPerson\DeleteList::create(['ids' => $organisationPersonIds]);
+        $result->merge($this->handleSideEffect($dto));
     }
 }
